+++ Client/GPlayer/CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
 
 Wed Feb 12 2025 23:22 - v0.0.2
+  - refactor: GPlayer.ui结构，播放列表最开始不显示
   - chore: 支持VS2017编译Release版本，去掉了gtest